LayoutTests/inspector/profiler/cpu-profiler-profiling-without-inspector.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js" type="application/x-javascript"></script>
<script type="application/x-javascript">

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

if (window.internals)
    internals.setJavaScriptProfilingEnabled(true);

// Here and below, the opening brace is on the same line as function declaration
// intendionally, as it compensates for differences between the ways JSC and V8
// use to report source lines for functions (V8 reports line of function(), JSC
// the line of the brace)
function sleep(milliseconds) {
    var spinUntil = Date.now() + milliseconds;
    while (Date.now() < spinUntil) {}
}

function functionB() {
    sleep(200);
}

function functionA() {
    sleep(200);
    functionB();
}

function startProfiling()
{
    console.profile("outer1");
    console.profile("inner1");  // [Chromium] Make sure we capture the current callstack.
    // workaround for http://crbug.com/164304
    setTimeout(doWork, 100);
}

function doWork() {
    functionA();
    console.profileEnd("outer1");
    console.profileEnd("inner1");
    printResult();
    if (window.testRunner)
         testRunner.notifyDone();
}

function startTest()
{
    startProfiling();
}

function printResult()
{
    var profiles = console.profiles;
    for (var i = 0; i < profiles.length; ++i) {
        var profile = profiles[i];
        if (profile.title !== "inner1")
            continue;
        dumpNode(profile.head)
    }
}

var functionsWhiteList = {
    "sleep": 1,
    "functionA": 1,
    "functionB": 1,
    "doWork": 1
};

function dumpNode(node, prefix)
{
    if (!node.visible)
        return;
    prefix = prefix || "";
    
    const roundingThreshold = 100; // Time should be good within 100ms
    var time = Math.round(node.totalTime / roundingThreshold) * roundingThreshold;
    if (time && functionsWhiteList[node.functionName])
        output(prefix + node.functionName + " (line " + node.lineNumber + "): ~" + time +"ms");

    var children = node.children();
    for (var i = 0; i < children.length; ++i)
        dumpNode(children[i], prefix + "    ");
}
</script>

</head>
<body onload="startTest()">
<p>
Tests that CPU profiling works.<br>
Doesn't open Inspector, uses <b>console.profile...</b>.

<div id="output"></div>
</p>
</body>
</html>
